+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Infinite loop when parsing MetricKit data (#3395)
 - Fix incorrect implementation in #3398 to work around a profiling crash (#3405)
+- Fix crash in SentryFramesTracker (#3424)
 
 ### Improvements